fixed bug where loading absolute offset vals were truncated

+++ bk_JUCE/BK/Source/BKUtilities.cpp
@@ -160,7 +160,8 @@ String offsetArrayToString(Array<float> arr)
     int key = 0;
     for (auto offset : arr)
     {
-        if (offset != 0.0)  s += String(key) + ":" + String((int)(offset*100.0f)) + " ";
+        //if (offset != 0.0)  s += String(key) + ":" + String((int)(offset*100.0f)) + " ";
+        if (offset != 0.0)  s += String(key) + ":" + String((offset*100.0f)) + " ";
         
         ++key;
     }
